refactor: 💡 fix code duplication
