+++ src/compliance/trusted-claim-issuers.controller.spec.ts
@@ -1,21 +1,17 @@
 import { DeepMocked } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
-import { BigNumber } from '@polymeshassociation/polymesh-sdk';
 import { TxTags } from '@polymeshassociation/polymesh-sdk/types';
 import { when } from 'jest-when';
 
-import { TransactionType } from '~/common/types';
 import { RemoveTrustedClaimIssuersDto } from '~/compliance/dto/remove-trusted-claim-issuers.dto';
 import { SetTrustedClaimIssuersDto } from '~/compliance/dto/set-trusted-claim-issuers.dto';
 import { TrustedClaimIssuersController } from '~/compliance/trusted-claim-issuers.controller';
 import { TrustedClaimIssuersService } from '~/compliance/trusted-claim-issuers.service';
-import { testValues } from '~/test-utils/consts';
-import { createMockTransactionResult, mockTrustedClaimIssuer } from '~/test-utils/mocks';
+import { createMockTxResult, mockTrustedClaimIssuer } from '~/test-utils/mocks';
 import { mockTrustedClaimIssuersServiceProvider } from '~/test-utils/service-mocks';
 
 describe('TrustedClaimIssuersController', () => {
   const mockParams = { ticker: 'TICKER' };
-  const { txResult } = testValues;
   let controller: TrustedClaimIssuersController;
   let mockService: DeepMocked<TrustedClaimIssuersService>;
 
@@ -55,17 +51,9 @@ describe('TrustedClaimIssuersController', () => {
 
   describe('setTrustedClaimIssuers', () => {
     it('should accept SetTrustedClaimIssuersDto and set Asset trusted claim issuers', async () => {
-      const transaction = {
-        blockHash: '0x1',
-        transactionHash: '0x2',
-        blockNumber: new BigNumber(1),
-        type: TransactionType.Single,
-        transactionTag: TxTags.complianceManager.AddDefaultTrustedClaimIssuer,
-      };
-      const testTxResult = createMockTransactionResult<void>({
-        ...txResult,
-        transactions: [transaction],
-      });
+      const testTxResult = createMockTxResult(
+        TxTags.complianceManager.AddDefaultTrustedClaimIssuer
+      );
       const mockPayload: SetTrustedClaimIssuersDto = {
         claimIssuers: [],
         signer: 'Alice',
@@ -83,17 +71,9 @@ describe('TrustedClaimIssuersController', () => {
 
   describe('addTrustedClaimIssuers', () => {
     it('should accept SetTrustedClaimIssuersDto and add Asset trusted claim issuers', async () => {
-      const transaction = {
-        blockHash: '0x1',
-        transactionHash: '0x2',
-        blockNumber: new BigNumber(1),
-        type: TransactionType.Single,
-        transactionTag: TxTags.complianceManager.AddDefaultTrustedClaimIssuer,
-      };
-      const testTxResult = createMockTransactionResult<void>({
-        ...txResult,
-        transactions: [transaction],
-      });
+      const testTxResult = createMockTxResult(
+        TxTags.complianceManager.AddDefaultTrustedClaimIssuer
+      );
       const mockPayload: SetTrustedClaimIssuersDto = {
         claimIssuers: [],
         signer: 'Alice',
@@ -111,17 +91,9 @@ describe('TrustedClaimIssuersController', () => {
 
   describe('removeTrustedClaimIssuers', () => {
     it('should accept RemoveTrustedClaimIssuersDto and remove trusted claim issuers for Asset', async () => {
-      const transaction = {
-        blockHash: '0x1',
-        transactionHash: '0x2',
-        blockNumber: new BigNumber(1),
-        type: TransactionType.Single,
-        transactionTag: TxTags.complianceManager.RemoveDefaultTrustedClaimIssuer,
-      };
-      const testTxResult = createMockTransactionResult<void>({
-        ...txResult,
-        transactions: [transaction],
-      });
+      const testTxResult = createMockTxResult(
+        TxTags.complianceManager.RemoveDefaultTrustedClaimIssuer
+      );
 
       const mockPayload: RemoveTrustedClaimIssuersDto = {
         claimIssuers: [],
