+++ src/test-utils/mocks.ts
@@ -7,6 +7,7 @@ import { SettlementResultEnum } from '@polymeshassociation/polymesh-sdk/middlewa
 import {
   Account,
   AuthorizationType,
+  ComplianceManagerTx,
   HistoricSettlement,
   MetadataEntry,
   MetadataType,
@@ -20,8 +21,12 @@ import {
 } from '@polymeshassociation/polymesh-sdk/types';
 import { Response } from 'express';
 
+import { TransactionType } from '~/common/types';
+import { ServiceReturn } from '~/common/utils';
+import { EventType } from '~/events/types';
+import { NotificationPayload } from '~/notifications/types';
 import { PolymeshService } from '~/polymesh/polymesh.service';
-import { testValues } from '~/test-utils/consts';
+import { testValues, txResult } from '~/test-utils/consts';
 import { TransactionResult } from '~/transactions/transactions.util';
 
 const { did } = testValues;
@@ -458,4 +463,23 @@ export function createMockResultSet<T extends any[]>(data: T): ResultSet<T> {
     next: '0',
     count: new BigNumber(data.length),
   };
+}
+
+export function createMockTxResult(
+  transactionTag: ComplianceManagerTx
+): TransactionResult<void> | ServiceReturn<void> | NotificationPayload<EventType> {
+  const transaction = {
+    blockHash: '0x1',
+    transactionHash: '0x2',
+    blockNumber: new BigNumber(1),
+    type: TransactionType.Single,
+    transactionTag: transactionTag,
+  };
+
+  const testTxResult = createMockTransactionResult<void>({
+    ...txResult,
+    transactions: [transaction],
+  });
+
+  return testTxResult;
 }
